+++ themes/shipsay/tpl_header.php
@@ -1,11 +1,13 @@
 <?php if (!defined('__ROOT_DIR__')) exit; ?>
 <?php
+$allbooks_url_safe = !empty($allbooks_url) ? $allbooks_url : '/sort/';
 $full_allbooks_url_safe = !empty($full_allbooks_url)
     ? $full_allbooks_url
-    : ('/quanben' . (isset($allbooks_url) ? $allbooks_url : '/sort/'));
+    : ('/quanben' . $allbooks_url_safe);
 $search_url_safe = function_exists('ss_search_url')
     ? ss_search_url()
     : ((isset($fake_search) && $fake_search) ? $fake_search : '/search/');
+$recentread_url_safe = !empty($fake_recentread) ? $fake_recentread : '/history.html';
 $search_placeholder = isset($search_placeholder) && $search_placeholder !== ''
     ? $search_placeholder
     : '输入书名/作者';
@@ -49,10 +51,10 @@ if (isset($fake_top) && $fake_top) {
 
         <div class="header_right">
             <a id="home" href="/"><i class="fa fa-home fa-lg"></i><br>首页</a>
-            <a href="<?=$allbooks_url?>"><i class="fa fa-book fa-lg"></i><br>书库</a>
+            <a href="<?=$allbooks_url_safe?>"><i class="fa fa-book fa-lg"></i><br>书库</a>
             <a href="<?=$full_allbooks_url_safe?>"><i class="fa fa-coffee fa-lg"></i><br>完本</a>
             <a href="<?=$rank_entry_safe?>"><i class="fa fa-bar-chart fa-lg"></i><br>排行</a>
-            <a href="<?=$fake_recentread?>" rel="nofollow"><i class="fa fa-history fa-lg"></i><br>足迹</a>
+            <a href="<?=$recentread_url_safe?>" rel="nofollow"><i class="fa fa-history fa-lg"></i><br>足迹</a>
         </div>
     </div>
 </header>